Moved car network components to the common_networks.c file
