--- Car/car_networks.h
@@ -1,3 +1,0 @@
-int connect_to_control_system();
-int sendMessage(int socketFd, const char *message);
-char *receive_floor(int serverFd);
