+++ common_networks.h
@@ -1,0 +1,2 @@
+int send_message(int socketFd, const char *message);
+char *receive_message(int socketFd);
